feat(fuyu): fubuki::fuyu
`fubuki::fuyu` provides an idiomatic C++ API (bindings) to the Vulkan API.
It provides objects with RAII semantics, that support `move` (but not copy), and are aware of how they were allocated. To a certain extend.

`fuyu`, as the rest of Fubuki, is a _functional_ API: the owning RAII objects convert to an equivalent representation in `fuyu::views`, trivially copiable objects that can be passed around in functions.